+++ test/integration/helpers.py
@@ -28,7 +28,6 @@ from tempfile import gettempdir, mkdtemp
 from typing import List, Optional, Union
 
 import backoff
-import pytest
 from b2sdk._v3.exception import BucketIdNotFound as v3BucketIdNotFound
 from b2sdk.v2 import (
     ALL_CAPABILITIES,
@@ -312,8 +311,8 @@ def run_command(
     environ['PYTHONPATH'] = '.'
     environ['PYTHONIOENCODING'] = 'utf-8'
     command = cmd.split(' ')
-    args = [str(arg) for arg in args]
-    command.extend(args or [])
+    args: List[str] = [str(arg) for arg in args] if args else []
+    command.extend(args)
 
     print('Running:', ' '.join(command))
 
@@ -530,10 +529,3 @@ def set_file_mod_time_millis(path: Union[str, Path], time):
 
 def random_hex(length):
     return ''.join(random.choice('0123456789abcdef') for _ in range(length))
-
-
-def skip_on_windows(*args, reason='Not supported on Windows', **kwargs):
-    return pytest.mark.skipif(
-        platform.system() == 'Windows',
-        reason=reason,
-    )(*args, **kwargs)