+++ test/unit/test_console_tool.py
@@ -20,7 +20,6 @@ from typing import List, Optional
 from b2sdk import v1
 from b2sdk.v2 import (
     ALL_CAPABILITIES,
-    REALM_URLS,
     B2Api,
     B2HttpApiConfig,
     ProgressReport,
@@ -247,7 +246,6 @@ class BaseConsoleToolTest(TestBase):
         bucket.upload(data, 'c/test.tsv')
 
 
-@mock.patch.dict(REALM_URLS, {'production': 'http://production.example.com'})
 class TestConsoleTool(BaseConsoleToolTest):
     def test_authorize_with_bad_key(self):
         expected_stdout = '''
@@ -2407,7 +2405,6 @@ class TestConsoleTool(BaseConsoleToolTest):
             assert parallel_strategy.max_streams == params['--max-download-streams-per-file']
 
 
-@mock.patch.dict(REALM_URLS, {'production': 'http://production.example.com'})
 class TestConsoleToolWithV1(BaseConsoleToolTest):
     """These tests use v1 interface to perform various setups before running CLI commands"""
 
@@ -2483,7 +2480,6 @@ class TestConsoleToolWithV1(BaseConsoleToolTest):
         self._run_command(['list-unfinished-large-files', 'my-bucket'], expected_stdout, '', 0)
 
 
-@mock.patch.dict(REALM_URLS, {'production': 'http://production.example.com'})
 class TestRmConsoleTool(BaseConsoleToolTest):
     """
     These tests replace default progress reporter of Rm class